Trait substrate_consensus_aura_primitives::AuraApi

pub trait AuraApi<Block: BlockT, AuthorityId: Codec>: Core<Block> {
    fn slot_duration(&self, at: &BlockId<Block>) -> Result<u64, Error> { ... }

    fn slot_duration_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, Error> { ... }

    fn authorities(
        &self, 
        at: &BlockId<Block>
    ) -> Result<Vec<AuthorityId>, Error> { ... }

    fn authorities_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Vec<AuthorityId>, Error> { ... }
}

API necessary for block authorship with aura.

Provided methods

fn slot_duration(&self, at: &BlockId<Block>) -> Result<u64, Error>

Return the slot duration in seconds for Aura. Currently, only the value provided by this type at genesis will be used.

Dynamic slot duration may be supported in the future.

fn slot_duration_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, Error>

Return the slot duration in seconds for Aura. Currently, only the value provided by this type at genesis will be used.

Dynamic slot duration may be supported in the future.

fn authorities(&self, at: &BlockId<Block>) -> Result<Vec<AuthorityId>, Error>

fn authorities_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<Vec<AuthorityId>, Error>

Loading content...